feat(grpc-xds): Add code for workshop
Adds sample Go and Java implementations of xDS control planes for
gRPC services running on Kubernetes, as well as accompanying sample
applications that can run as xDS-enabled gRPC servers.

These sample implementations are not recommended for production
deployments. They are intended to provide a practical understanding
of xDS, beyond the API specifications and gRFCs.