feat(grpc-xds): Control plane mTLS for Go
Implements mTLS for connections from xDS clients to the xDS control
plane management server as per gRFC A65.

gRPC-Java as of v1.61.0 does not yet support "tls" type channel
credentials in the gRPC xDS bootstrap file, so control plane mTLS
is disabled by default for the Java implementation here.

On the control plane side, the feature is controlled by flags in
the `xds_features.yaml` file.

On the client side, there is a Kubernetes resource patch for the
Java deployment that swaps `"type": "tls"` for
`"type": "insecure"`.