feat(grpc-xds): Multiple Kubernetes clusters in different regions
Create two Kubernetes clusters. They are in the same VPC network,
but they can be in different cloud provider regions.

Deploy the control plane to one Kubernetes cluster and the sample
gRPC applcation to both Kubernetes clusters.

greeter-intermediary in one zone use a greeter-leaf server in the
same zone if available. If not, it can fall back to a greeter-leaf
server in a different zone of the same region. And if none are
available and healthy there, it can then automatically use a
greeter-leaf server in another Kubernetes cluster, which could
be in another region, and even in another cloud provider.
Assuming there is connectivity between the providers of course,
e.g., using VPN.