feat(grpc-xds): Locality-based load balancing
Assign priorities to localities in EDS ClusterLoadAssignment
resources, based on the zone of the requesting xDS node.

xDS clients are grouped by zone, where all clients in a zone get
the same xDS resource snapshot. The priorities in the EDS resource
are determined by comparing the xDS client zone with the endpoint
zone.

The priority order in this implementation is defined as follows:

1. Same zone
2. Same region (e.g., us-west1-a and us-west1-b are both us-west1-*)
3. Same super-region (e.g., us-west1-a and us-west2-a are both us-west*)
4. Same multi-region (e.g., us-east1-a and us-west1-a are both us*)
5. Other - no match